Move error to apiHandlers
